--Fixed warnings --Added printfs
git-svn-id: file:///home/skmp/Desktop/public_svn/nullDC@755 d5fdb02f-ce0d-0410-91b1-fe443107cc8c

+++ trunk/nulldc/plugins/ImgReader/mds.cpp
@@ -178,7 +178,7 @@ bool mds_init(wchar* file)
 		if (wcslen(file)>4)
 		{
 			wcscpy(&fn[0],file);
-			int len=wcslen(fn);
+			size_t len=wcslen(fn);
 			wcscpy(&fn[len-4],L".mdf");
 			
 			fp_mdf=_tfopen(fn,L"rb");
